Move Ledger handle back into hydra-node
The ledger and its Cardano/Simple instances are only used in hydra-node.

Also, all usage of SimpleTx in hydra-tx can be avoided as its
cardano-specific things after all.
