+++ hydra-node/bench/micro-bench/Main.hs
@@ -14,7 +14,7 @@ import Hydra.Cardano.Api (
  )
 import Hydra.Chain.ChainState (ChainSlot (ChainSlot))
 import Hydra.Ledger.Cardano (Tx, cardanoLedger, genFixedSizeSequenceOfSimplePaymentTransactions)
-import Hydra.Ledger.Ledger (Ledger (applyTransactions), ValidationError)
+import Hydra.Ledger (Ledger (applyTransactions), ValidationError)
 import Test.Hydra.Tx.Fixture (defaultGlobals, defaultLedgerEnv)
 import Test.QuickCheck (generate)
