+++ hydra-node/test/Hydra/Model/MockChain.hs
@@ -68,7 +68,7 @@ import Hydra.HeadLogic (
  )
 import Hydra.Ledger.Cardano (adjustUTxO, fromChainSlot)
 import Hydra.Ledger.Cardano.Evaluate (eraHistoryWithoutHorizon, evaluateTx)
-import Hydra.Ledger.Ledger (Ledger (..), ValidationError (..), collectTransactions)
+import Hydra.Ledger (Ledger (..), ValidationError (..), collectTransactions)
 import Hydra.Logging (Tracer)
 import Hydra.Model.Payment (CardanoSigningKey (..))
 import Hydra.Network (Network (..))
